// app/backend/src/database/models/LeaderBoardData.ts
import ILearderBoard from '../../interfaces/ILeaderboard';

export default class LeaderBoardData implements ILearderBoard {
  name: string;
  totalPoints: number;
  totalGames: number;
  totalVictories: number;
  totalDraws: number;
  totalLosses: number;
  goalsFavor: number;
  goalsOwn: number;
  goalsBalance: number;
  efficiency: string;
}